refactor: electron git
