+++ src/components/organisms_new/ResourceSetSelector/GitSelect/GitBranchSelect.tsx
@@ -29,7 +29,7 @@ const GitBranchSelect: React.FC<IProps> = ({side}) => {
   return (
     <S.SelectColor>
       <Select
-        defaultOpen
+        defaultOpen={!resourceSet}
         onChange={handleSelect}
         placeholder="Choose Branch..."
         value={currentGitBranch?.name}
