+++ src/components/organisms_new/ResourceSetSelector/HelmSelect/HelmChartSelect.tsx
@@ -29,7 +29,7 @@ export const HelmChartSelect: React.FC<Props> = ({side}) => {
   return (
     <S.SelectColor>
       <Select
-        defaultOpen
+        defaultOpen={!resourceSet}
         onChange={handleSelect}
         placeholder="Choose Chart…"
         value={currentHelmChart?.id}
