+++ src/components/organisms_new/ResourceSetSelector/KustomizeSelect.tsx
@@ -29,7 +29,7 @@ export const KustomizeSelect: React.FC<Props> = ({side}) => {
   return (
     <S.SelectColor>
       <Select
-        defaultOpen
+        defaultOpen={!resourceSet}
         onChange={id => handleSelect(id as string)}
         placeholder="Choose Kustomization…"
         value={currentKustomization?.id}
